Initial commit of Cloud Native Data Plane
A new cloud native centric data plane framework is needed to
meet the needs of high-performance, cloud native packet processing
applications and the design principles of this paradigm.

CNDP is a collection of user space libraries for accelerating packet
processing for cloud applications. It aims to provide better
performance than that of standard network socket interfaces
using an I/O layer primarily built on AF_XDP, an interface that
delivers packets directly to user space, bypassing the
kernel networking stack.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>
Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>
Signed-off-by: Maryam Tahhan <maryam.tahhan@intel.com>
Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Signed-off-by: Dakshina Ilangovan <dakshina.ilangovan@intel.com>
Signed-off-by: Edwin Verplanke <edwin.verplanke@intel.com>
Signed-off-by: Elza Mathew <elza.mathew@intel.com>
Signed-off-by: Magnus Karlsson <magnus.karlsson@intel.com>
Signed-off-by: Manoj Gopalakrishnan <manoj.gopalakrishnan@intel.com>
Signed-off-by: Sushma Sitaram <sushma.sitaram@intel.com>