go: rework directory layout
Rework the lang/go directory layout, remove some old components and
rename the binding layer to cne instead of cndpgo. The rename of the modules
to be inline with standard go applications on github.
Rerun go mod tidy making sure we have go.mod files correct for the
new layout. The go.sum files should be included in the git repo.

I have ran both bindings/examples/fwd and tool/cmon to verify they work need
to verify metrics and prometheus work in a container or k8s. The new bindings
directory will contain other binding layers i.e. hyperscan, multi-buffer IPsec, ...

lang/go
├── bindings
│   ├── cne        # renamed from cndpgo to cne as we should not use cndp.
│   └── examples
│       └── fwd
├── stats
│   ├── metrics
│   └── prometheus
└── tools
    ├── cmon
    └── pkgs
        ├── asciichart
        │   ├── cmd
        │   └── examples
        ├── colorize
        ├── etimers
        ├── graphdata
        ├── hexdump
        ├── irq
        ├── taborder
        ├── ttylog
        └── utils

One areas that was addressed was the naming of modules to align with
the public github.com directory structure. The github.com account is
call github.com/CloudNativeDataPlane, which means we need to rename
the modules. The CNDP repo on github.com will be called 'cndp' i.e.
 --> cndp.org/cmon is now github.com/CloudNativeDataPlane/cndp/lang/go/tools/cmon

Modified lang/go/stats/metrics to not require ttylog, which makes this
module not dependent on lang/go/tools/pkgs/ttylog.

For now we can leave cmom in the CNDP repo, but it may be moved to a new repo
in the future. I would like to make better integrated into CNDP as a tool to
help configure and understand the configuration of CNDP application and the system.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>