vec:make vec lists dynamic
Make the vec lists more dynamic instead of static sizes. Part of this change
is because of the CNET/TCP update comming, but also makes the code simpler.

With this update the old vec_*_ptr routines are not required as the macros
detect the correct size and type of vector lists.

The number of files changed is fair size, but most of the changes are minor
renaming of macros.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>