tcp:add tcp support to CNET
Update and rework the TCP support for CNET using a set of graph nodes to
manage the TCP protocol. The TCP protocol support can be enabled via the
meson_options.txt file. Currently, the TCP support experimental. I attempted
to reduce the number of changes in one PR, but that seemed to be very difficult
to create PRs which would be functional.

A number of changes to the code affects a number of files, which include:
  - remove the TCP send graph node as it is not required now.
  - create a chnl_priv.h file to hold internal items.
    - adding the new header touched a number of files to include the file.
  - change the channel APIs to use the new channel descriptor index value.
    - the change from pointer to index affected a number of files more below.
  - the cnet-graph and cnet-quic examples needed to change to support TCP and/or
    to use the new callback support with the channel descriptor changes.
  - add more doxygen documentation for APIs which did not have comments.
  - remove a number of channel options as they are not required now.
  - add a number of options to the meson_options.txt file to help configure
    and enable some TCP debug code. The change enable removing some defines
    in the code to now be configurable.
  - reworked some of the locking in channel and stack code, which helped to
    fix a number of klocwork issues.
  - ran klocwork and fixed a number of issues

Testing the TCP protocol using applications i.e socat, telnet, ...
to verify the protocol and data management is correct.

The Channel APIs no longer use a 'struct chnl' pointer, replaced with a channel
descriptor index similar to file descriptors. The change affected many files to
now support using a descriptor value instead of a pointer. Internally the CNET
stack will use a 'struct chnl *' value, but public interfaces must use the
channel descriptor index value.

The CNET stack uses a callback method to keep packet processing to the same thread
as the graph node handling the packets. Using a callback helps to avoid locks and
other complexity issues with the CNET stack.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>