tcp:add tcp support to CNET
Update and rework the TCP support for CNET using a set of graph nodes to
manage the TCP protocol. The TCP protocol support can be enabled via the
meson_options.txt file.

Testing the TCP protocol using wireguard and other applications i.e socat, telnet, ...
to verify the protocol and data managment is correct.

In this change a few examples needed to be update to support TCP and the changes to
the channel APIs. One change to the channel API is when the application is called
back to process Rx packets is the callback does not pass mbufs back in the call.
The packets are now placed on a receive queue and a call to chnl_recv() needs to
be done to get packets.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>