build:update deprecated functions
Starting in meson build 0.56 meson.build_root() and meson.source_root() are
being deprecated. This was found with Ubuntu 22.04 and meson version 0.61.2

Meson build reports the following.
NOTICE: Future-deprecated features used:
 * 0.56.0: {'meson.build_root', 'meson.source_root'}

It appears that meson.current_source_root() and meson.current_build_root() is
not deprecated and we can convert to using these.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>