testcne: add ability to mark test as skipped
In addition to a test failing or passing, it can be useful to mark a
test as skipped, in case it cannot run due to lack of privilege or
for some reason that can only be determined after running the test.

To distinguish these from failures (or worse, passing tests) the test
can be skipped. When run with `meson test -C builddir`, meson will
detect the skipped test, and still return without error.

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>