jcfg: fix lport group mapping onto same thread
The lport group processing fails for two groups mapping to a single
thread. This is because the lport list size was not updated after
reallocating space for the new lports in the first lport group. The
second lport group saw the old size, so reallocated a smaller array than
is necessary to store the existing plus new lports in the per-thread
array.

Fixes: c83ca9e5ea ("Initial commit of Cloud Native Data Plane")

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>