jcfg: fix lport group umem creation
A default umem is created for all logical ports belonging to a logical
port group which does not otherwise define a umem. Previously, a umem
was created for each lport group, not a single one for all lports. To
avoid a packet copy when transmitting packets received on a lport in a
different group, all the groups should use a common umem.

Signed-off-by: Jeff Shaw <jeffrey.b.shaw@intel.com>