ring:add a cne_ring_init to use external memory
Add a new API cne_ring_init() to allow the caller to pass in the memory
to build the ring into for allowing the caller to supply the memory region.

The new API becomes useful when we need to share memory or the application needs
to build a ring into a already allocated memory region.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>