add pktmbuf external metadata
The metadata was contained in the headroom of the pktmbuf and it is
limited in size to max headroom size normally 128 bytes.
The headroom size is also used for adding protcol headers and if the
room is consumed by the metadata then headers can not be added.

Created a new API pktmbuf_pool_cfg_create() to be able to create a
external metadata buffer array indexed by the pktmbuf_t.metadata_idx
value. The new metadata array of buffers must be a multiple of a
cacheline (64 bytes) and the number of buffers will match the number of
pktmbuf_t buffers in the mempool.

If the new API pktmbuf_pool_cfg_create() is not used the metadata buffer
will be part of the headroom in the pktmbuf buffer.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>