usrtools: utility to send pinned xsk_map to cndp
This commit introduces a utility that will allow a user to send
a pinned xsk_map to a CNDP application. It allows the user to
write their own BPF programs and load them with their preferred
BPF program loader. The only constraint is that they must pin
their XSK_MAP.

Signed-off-by: Maryam Tahhan <mtahhan@redhat.com>