fix pktdev_close to release internal lport data
When pktdev_close() was called the code would close the lport
but it needed to release the lport by calling pktdev_release_port().

Cleanup the a PMD probe and remove routines. In the probe routine we
now store the struct pktdev_driver pointer in the struct cne_pktdev structure
to allow for a cleaner remove or close of an lport. Remove pktdev_create_done()
routine and move the logic into the pktdev_port_setup() routine.

This would allow the Go code to be able to cleanup correctly and the PMDs
were checking if dev == NULL before calling pktdev_release_port() which is
not required.

Remove the PKTDEV_LOG() macro and use CNE_XXX macros instead.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>