simulate xdp port stats reset
When the xskdev_stats_reset() is called the lport_stats_t is set to zero,
but the XDP stats are not reset as they are handled by the kernel. This
means the counter we display like imissed or inpacket do not appear to
be reset to zero. We can not reset the XDP stats so keep a copy on init
of the port and after a reset of the stats on a lport.

Add some logic to simulate the XDP stats are reset by keeping a original
set of stats from the XDP counters and substract the difference to
get the correct value after a reset call.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>