only enable gofmt for golangci-linter
golangci-linter runs a set of linters by default if we use the
golangci-linter through super-linter. This was causing some false
positives to be seen. For example, in this PR -
https://github.com/CloudNativeDataPlane/cndp/runs/7883419903?check_suite_focus=true#step:4:474
Adding a config file for golangci-linter so that only the gofmt linter is enabled for now and
no false positives are seen. Other linters are disabled.

Signed-off-by: Elza Mathew <elza.mathew@intel.com>