add idle manager library and test-cne test
The current design in CNDP is 100% polling of the receive ring in AF_XDP
or PMD which means burning core power. The patch adds two parameters to
the jsonc file to control polling idle_timeout and intr_timeout
in thread section.

The idlemgr library handles managing epoll and calling epoll_wait when the
idle_timeout value has been met, then it calls epoll_wait(). The idlemgr
operates on threads, using file descriptors as the method to wakeup a sleeping
thread. The caller needs to add file descriptors to the idlemgr instance and
then call idlemgr_process() with a flag if the thread thinks it is idle or
not.

The first one gives the number of milliseconds to wait for the RX ring to
be idle, then call epoll() with a timeout of intr_timeout. When no Rx traffic
for a given time the idlemgr will call epoll(), which reduces the lcore load to
effectively zero and only waking up when packets arrive or a timeout occurs.

In testing of performance it appears to be very little impact when interrupt
mode is enabled compared to when it is not enabled. Added some counters to help
determine how the new mode is operating.

The idle_timeout value in the jsonc file for given thread is how this feature
is controlled. If not defined or set to zero interrupt mode is disabled. When
set to a non zero value will enable interrupt mode. The intr_timeout value
is only used if idle_timeout is non-zero and will be used in the poll() call
as the timeout value. Each of these values are in milliseconds.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>