Add jcfg support for idlemgr
The idlemgr timeout values enabled in the jsonc file in the thread section.
This commit will add the parsing of the two variables idle_timeout and
intr_timeout and place them in the jcfg thread structure.

This commit also fixes a duplicate define and missing error. The jcfg_lport.c
was updated to test the busy timeout and busy budget parsing to look just
like the jcfg_lport_group.c code.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>