fix the cne_register_test()
The cne_register_test() function must be run inside a new pthread or
it will corrupt the per_thread__cne variable of the current thread.

The problem really appears when the cne_register() occurs
as it replaces the threads per_thread__cne variable with the new
registered value corrupting the main thread value.

The cne_id() would detect per_thread__cne as null and return -1,
but it makes more sense to print a message and exit error as the
system is corrupted.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>