fix up registering go threads
Each Go thread wanting to call CNDP APIs needs to do a RegisterThread()
call to allocate a unique identifier used by mempool cache support.
The call also does a LockOSThread to pin the Go thread to the same
system thread as thread local storage is used in CNDP.

The RegisterThread() and UnregisterThread() routines did not need
the System receiver pointer to be passed into the routines. This
change removes the need to call these routines with a System pointer.

Use cne.RegisterThread("foo") and cne.UnregisterThread(int).

The RegisterThread and UnregisterThread routine used a mutex to
lock access to the cne_register(), but cne_register() now has
locks to protect the register and unregister calls.

The lport_test.go and system_test.go have been changed to move the
t.Run() calls to include the lport for loop instead of calling register
for each lport is not required and the RegisterThread()/UnregisterThread()
calls are only done once per t.Run() call. The lport_test.go benchmark
tests were removed.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>