fix up performance drop for FQ
The FQ routine fq_add() was allocating mbufs and freeing them when the
FQ was full, this caused a performance drop. The code will now determine
if the FQ needs mbufs and how many before allocating the mbufs from
the pktmbuf pool.

During testing of tx only cndpfwd and the idlemgr was enable causing
the thread to sleep due to no RX traffic. Modified the txonly routine
to return 1 to make it look like the port was active to idlemgr.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>