xskdev:force zero copy mode
When the driver did not support zero-copy mode it would default
to copy mode, but we did not force zero-copy mode. The performance
would drop and we did not report an error message to alert the
operator.

Force zero copy mode unless asked to do copy mode i.e., skb_mode.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>