fix static/shared executables to include pmds
With the new changes to building static PMDs the examples no longer
run correctly as the static PMDs are not linked to the images.

Need to add a linker option to the executable() function to include
the new static PMDs into the image. Also with static PMDs we can remove
the reference of the PMDs in the deps variables. If the application needs
headers from the PMDs then it must be included, but we do not need
anything from the PMDs to run the examples.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>