Fix build /tmp conflict on multi-user system
This fixes a problem where two files are placed in /tmp that are owned
by the user. The next user that tries to build on the system fails,
because those files aren't writable to them.

This patch puts the /tmp files in a separate randomly generated temp
directory in /tmp to avoid this conflict.

Signed-off-by: Stephen Weeks <stephen.weeks@intel.com>
Signed-off-by: Byron Marohn <byron.marohn@intel.com>
Signed-off-by: Matthew Leon <matthew.leon@intel.com>
Signed-off-by: Edmund Leemhuis <edmund.leemhuis@intel.com>