Add L3 forwarding to cndpfwd example
This commit enhances the cndpfwd example application by adding a new
`l3-fwd` mode. In this mode, incoming packets are forwarded based on
a FIB defined in `fwd.jsonc`, in the `options:l3fwd-fib-rules`
section. Each FIB entry consists of an IP/prefix to match, a
destination MAC for the next hop, and a TX port to use.

Targets #231.

Co-authored-by: Steve Miller <steve.miller@intel.com>
Signed-off-by: Andrew Capatina <andrew.capatina@intel.com>
Signed-off-by: Steve Miller <steve.miller@intel.com>