add hyperscan forwarding support
Add a simple forwarding mode using Hyperscan regex expressions to
cndpfwd application. The example can be expanded to be a DDOS or DPI or
a load balancer using Hyperscan. The example needs Hyperscan
installed to work.

The cndpfwd application in Hyperscan mode will drop or forward packets
to a different port. The Hyperscan string format is
<ID>:/<regex>/<flags> flags are the standard Hyperscan flags.

The ID number can be in hex or decimal, but hex 0x00000000 is easier
to use as the value is encoded with the port ID value in bits 16-23
, which is a value 0-N N being lport ID. The bits 24-31 can be used for
flags and the only flag today is 0x80, which means the packets that
match this expression should be dropped. The lower 16 bits is a
unique ID value for each expression and is not used by the application.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>