fix dependbot issues
The BCP 47 tag parser has quadratic time complexity due to inherent
aspects of its design. Since the parser is, by design, exposed to
untrusted user input, this can be leveraged to force a program to
consume significant time parsing Accept-Language headers. The parser
cannot be easily rewritten to fix this behavior for various reasons.
Instead the solution implemented in this CL is to limit the total
complexity of tags passed into ParseAcceptLanguage by limiting the
number of dashes in the string to 1000. This should be more than
enough for the majority of real world use cases, where the number
of tags being sent is likely to be in the single digits.

Dependbot issue #5, #6, #7

Signed-off-by: Keith Wiles <keith.wiles@intel.com>