tcp-chnl: add TCP connection established callback for connect and accept
The commit adds calls the TCP connection callback for clients when they
establish a connection i.e. the connection is ready to send and receive
data.

To simplify things, `CHNL_TCP_ACCEPT_TYPE` is renamed to
`CHNL_TCP_ESTABLISHED_TYPE`. The callback type is used for both
server's `accept` and client's `connect`. In case of accept, the
listening channel descriptor is passed to the callback. In case of
connect, the client channel descriptor is passed.

Signed-off-by: Jalal Mostafa <jalal.a.mostapha@gmail.com>